VietTQ [Refactor] Views
