+++ front-end/package.json
@@ -26,6 +26,7 @@
     "react": "^17.0.1",
     "react-chartist": "^0.14.4",
     "react-dom": "^17.0.1",
+    "react-froala-wysiwyg": "^3.2.6-1",
     "react-google-maps": "^9.4.5",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
