+++ front-end/src/views/Admin/PostManagement/PostManagement.tsx
@@ -1,0 +1,9 @@
+const PostManagement = (props: any) => {
+    return (
+        <div>
+            POST MANAGEMENT
+        </div>
+    )
+}
+
+export default PostManagement;
