+++ front-end/src/views/Admin/RTLPage/RTLPage.tsx
@@ -19,30 +19,30 @@ import BugReport from '@material-ui/icons/BugReport';
 import Code from '@material-ui/icons/Code';
 import Cloud from '@material-ui/icons/Cloud';
 // core components
-import Button from '../../components/MaterialCommon/CustomButtons/Button';
-import GridItem from '../../components/MaterialCommon/Grid/GridItem';
-import GridContainer from '../../components/MaterialCommon/Grid/GridContainer';
-import Table from '../../components/MaterialCommon/Table/Table';
-import Tasks from '../../components/MaterialCommon/Tasks/Tasks';
-import CustomTabs from '../../components/MaterialCommon/CustomTabs/CustomTabs';
-import Danger from '../../components/MaterialCommon/Typography/Danger';
-import Card from '../../components/MaterialCommon/Card/Card';
-import CardHeader from '../../components/MaterialCommon/Card/CardHeader';
-import CardAvatar from '../../components/MaterialCommon/Card/CardAvatar';
-import CardIcon from '../../components/MaterialCommon/Card/CardIcon';
-import CardBody from '../../components/MaterialCommon/Card/CardBody';
-import CardFooter from '../../components/MaterialCommon/Card/CardFooter';
-import SnackbarContent from '../../components/MaterialCommon/Snackbar/SnackbarContent';
+import Button from 'components/MaterialCommon/CustomButtons/Button';
+import GridItem from 'components/MaterialCommon/Grid/GridItem';
+import GridContainer from 'components/MaterialCommon/Grid/GridContainer';
+import Table from 'components/MaterialCommon/Table/Table';
+import Tasks from 'components/MaterialCommon/Tasks/Tasks';
+import CustomTabs from 'components/MaterialCommon/CustomTabs/CustomTabs';
+import Danger from 'components/MaterialCommon/Typography/Danger';
+import Card from 'components/MaterialCommon/Card/Card';
+import CardHeader from 'components/MaterialCommon/Card/CardHeader';
+import CardAvatar from 'components/MaterialCommon/Card/CardAvatar';
+import CardIcon from 'components/MaterialCommon/Card/CardIcon';
+import CardBody from 'components/MaterialCommon/Card/CardBody';
+import CardFooter from 'components/MaterialCommon/Card/CardFooter';
+import SnackbarContent from 'components/MaterialCommon/Snackbar/SnackbarContent';
 
 import {
   dailySalesChart,
   emailsSubscriptionChart,
   completedTasksChart
-} from '../../variables/charts';
+} from 'variables/charts';
 
-import rtlStyle from '../../assets/jss/material-dashboard-react/views/rtlStyle';
+import rtlStyle from 'assets/jss/material-dashboard-react/views/rtlStyle';
 
-import avatar from '../../assets/img/faces/marc.jpg';
+import avatar from 'assets/img/faces/marc.jpg';
 
 let bugs = [
   'طراح گرافیک از این متن به عنوان عنصری از ترکیب بندی برای پر کردن؟',
